Change package name from "user.agent" to "useragent".
