+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/useragent/amqp/servicebus/ServiceBusSenderUserAgentTest.java
@@ -1,14 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.cloud.autoconfigure.user.agent.amqp.servicebus;
+package com.azure.spring.cloud.autoconfigure.useragent.amqp.servicebus;
 
 import com.azure.core.util.ClientOptions;
 import com.azure.messaging.servicebus.ServiceBusClientBuilder;
 import com.azure.spring.cloud.autoconfigure.properties.AzureGlobalProperties;
 import com.azure.spring.cloud.autoconfigure.servicebus.AzureServiceBusAutoConfiguration;
-import com.azure.spring.cloud.autoconfigure.user.agent.util.UserAgentTestUtil;
-import com.azure.spring.service.servicebus.factory.ServiceBusSessionReceiverClientBuilderFactory;
+import com.azure.spring.cloud.autoconfigure.useragent.util.UserAgentTestUtil;
+import com.azure.spring.service.servicebus.factory.ServiceBusSenderClientBuilderFactory;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
@@ -17,15 +17,14 @@ import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import static com.azure.spring.core.AzureSpringIdentifier.AZURE_SPRING_SERVICE_BUS;
 import static org.assertj.core.api.Assertions.assertThat;
 
-class ServiceBusSessionReceiverUserAgentTest {
+class ServiceBusSenderUserAgentTest {
 
     @Test
     void shareServiceBusClientBuilderUserAgentTest() {
         userAgentTest(
             "spring.cloud.azure.servicebus.connection-string=Endpoint=sb://sample.servicebus.windows.net/;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=key",
-            "spring.cloud.azure.servicebus.consumer.entity-name=sample",
-            "spring.cloud.azure.servicebus.consumer.entity-type=QUEUE",
-            "spring.cloud.azure.servicebus.consumer.session-enabled=true"
+            "spring.cloud.azure.servicebus.producer.entity-name=sample",
+            "spring.cloud.azure.servicebus.producer.entity-type=QUEUE"
         );
     }
 
@@ -33,10 +32,9 @@ class ServiceBusSessionReceiverUserAgentTest {
     void notShareServiceBusClientBuilderUserAgentTest() {
         userAgentTest(
             "spring.cloud.azure.servicebus.connection-string=Endpoint=sb://sample.servicebus.windows.net/;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=key",
-            "spring.cloud.azure.servicebus.consumer.entity-name=sample",
-            "spring.cloud.azure.servicebus.consumer.entity-type=QUEUE",
-            "spring.cloud.azure.servicebus.consumer.session-enabled=true",
-            "spring.cloud.azure.servicebus.consumer.namespace=sample"
+            "spring.cloud.azure.servicebus.producer.entity-name=sample",
+            "spring.cloud.azure.servicebus.producer.entity-type=QUEUE",
+            "spring.cloud.azure.servicebus.producer.namespace=sample"
         );
     }
 
@@ -46,11 +44,11 @@ class ServiceBusSessionReceiverUserAgentTest {
             .withBean(AzureGlobalProperties.class, AzureGlobalProperties::new)
             .withPropertyValues(propertyValues)
             .run(context -> {
-                assertThat(context).hasSingleBean(ServiceBusSessionReceiverClientBuilderFactory.class);
-                assertThat(context).hasSingleBean(ServiceBusClientBuilder.ServiceBusSessionReceiverClientBuilder.class);
+                assertThat(context).hasSingleBean(ServiceBusSenderClientBuilderFactory.class);
+                assertThat(context).hasSingleBean(ServiceBusClientBuilder.ServiceBusSenderClientBuilder.class);
 
-                ServiceBusClientBuilder.ServiceBusSessionReceiverClientBuilder receiverClientBuilder = context.getBean(ServiceBusClientBuilder.ServiceBusSessionReceiverClientBuilder.class);
-                ServiceBusClientBuilder builder = (ServiceBusClientBuilder) UserAgentTestUtil.getPrivateFieldValue(ServiceBusClientBuilder.ServiceBusSessionReceiverClientBuilder.class, "this$0", receiverClientBuilder);
+                ServiceBusClientBuilder.ServiceBusSenderClientBuilder serviceBusSenderClientBuilder = context.getBean(ServiceBusClientBuilder.ServiceBusSenderClientBuilder.class);
+                ServiceBusClientBuilder builder = (ServiceBusClientBuilder) UserAgentTestUtil.getPrivateFieldValue(ServiceBusClientBuilder.ServiceBusSenderClientBuilder.class, "this$0", serviceBusSenderClientBuilder);
                 ClientOptions options = (ClientOptions) UserAgentTestUtil.getPrivateFieldValue(ServiceBusClientBuilder.class, "clientOptions", builder);
                 Assertions.assertNotNull(options);
                 Assertions.assertEquals(AZURE_SPRING_SERVICE_BUS, options.getApplicationId());
