+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/useragent/amqp/servicebus/ServiceBusSessionReceiverUserAgentTest.java
@@ -1,14 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.cloud.autoconfigure.user.agent.amqp.servicebus;
+package com.azure.spring.cloud.autoconfigure.useragent.amqp.servicebus;
 
 import com.azure.core.util.ClientOptions;
 import com.azure.messaging.servicebus.ServiceBusClientBuilder;
 import com.azure.spring.cloud.autoconfigure.properties.AzureGlobalProperties;
 import com.azure.spring.cloud.autoconfigure.servicebus.AzureServiceBusAutoConfiguration;
-import com.azure.spring.cloud.autoconfigure.user.agent.util.UserAgentTestUtil;
-import com.azure.spring.service.servicebus.factory.ServiceBusReceiverClientBuilderFactory;
+import com.azure.spring.cloud.autoconfigure.useragent.util.UserAgentTestUtil;
+import com.azure.spring.service.servicebus.factory.ServiceBusSessionReceiverClientBuilderFactory;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
@@ -17,14 +17,15 @@ import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import static com.azure.spring.core.AzureSpringIdentifier.AZURE_SPRING_SERVICE_BUS;
 import static org.assertj.core.api.Assertions.assertThat;
 
-class ServiceBusReceiverUserAgentTest {
+class ServiceBusSessionReceiverUserAgentTest {
 
     @Test
     void shareServiceBusClientBuilderUserAgentTest() {
         userAgentTest(
             "spring.cloud.azure.servicebus.connection-string=Endpoint=sb://sample.servicebus.windows.net/;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=key",
             "spring.cloud.azure.servicebus.consumer.entity-name=sample",
-            "spring.cloud.azure.servicebus.consumer.entity-type=QUEUE"
+            "spring.cloud.azure.servicebus.consumer.entity-type=QUEUE",
+            "spring.cloud.azure.servicebus.consumer.session-enabled=true"
         );
     }
 
@@ -34,6 +35,7 @@ class ServiceBusReceiverUserAgentTest {
             "spring.cloud.azure.servicebus.connection-string=Endpoint=sb://sample.servicebus.windows.net/;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=key",
             "spring.cloud.azure.servicebus.consumer.entity-name=sample",
             "spring.cloud.azure.servicebus.consumer.entity-type=QUEUE",
+            "spring.cloud.azure.servicebus.consumer.session-enabled=true",
             "spring.cloud.azure.servicebus.consumer.namespace=sample"
         );
     }
@@ -44,11 +46,11 @@ class ServiceBusReceiverUserAgentTest {
             .withBean(AzureGlobalProperties.class, AzureGlobalProperties::new)
             .withPropertyValues(propertyValues)
             .run(context -> {
-                assertThat(context).hasSingleBean(ServiceBusReceiverClientBuilderFactory.class);
-                assertThat(context).hasSingleBean(ServiceBusClientBuilder.ServiceBusReceiverClientBuilder.class);
+                assertThat(context).hasSingleBean(ServiceBusSessionReceiverClientBuilderFactory.class);
+                assertThat(context).hasSingleBean(ServiceBusClientBuilder.ServiceBusSessionReceiverClientBuilder.class);
 
-                ServiceBusClientBuilder.ServiceBusReceiverClientBuilder receiverClientBuilder = context.getBean(ServiceBusClientBuilder.ServiceBusReceiverClientBuilder.class);
-                ServiceBusClientBuilder builder = (ServiceBusClientBuilder) UserAgentTestUtil.getPrivateFieldValue(ServiceBusClientBuilder.ServiceBusReceiverClientBuilder.class, "this$0", receiverClientBuilder);
+                ServiceBusClientBuilder.ServiceBusSessionReceiverClientBuilder receiverClientBuilder = context.getBean(ServiceBusClientBuilder.ServiceBusSessionReceiverClientBuilder.class);
+                ServiceBusClientBuilder builder = (ServiceBusClientBuilder) UserAgentTestUtil.getPrivateFieldValue(ServiceBusClientBuilder.ServiceBusSessionReceiverClientBuilder.class, "this$0", receiverClientBuilder);
                 ClientOptions options = (ClientOptions) UserAgentTestUtil.getPrivateFieldValue(ServiceBusClientBuilder.class, "clientOptions", builder);
                 Assertions.assertNotNull(options);
                 Assertions.assertEquals(AZURE_SPRING_SERVICE_BUS, options.getApplicationId());
